Created Card component & fixed styling of BM Item
